+++ packages/nimble-components/src/label-provider/rich-text/tests/label-provider-rich-text.stories.ts
@@ -5,7 +5,6 @@ import {
 } from '../../base/tests/label-provider-stories-utils';
 import { labelProviderRichTextTag } from '..';
 import * as labelTokensNamespace from '../label-tokens';
-import { removePrefixAndCamelCase } from '../../base/tests/label-name-utils';
 
 const metadata = {
     ...labelProviderMetadata,
@@ -18,7 +17,6 @@ export const richTextLabelProvider: StoryObj<LabelProviderArgs> = {
     args: {
         labelProviderTag: labelProviderRichTextTag,
         labelTokens: Object.entries(labelTokensNamespace),
-        prefixSubstring: 'richText',
-        removeNamePrefix: removePrefixAndCamelCase
+        prefixSubstring: 'richText'
     }
 };