+++ packages/nimble-components/src/label-provider/table/tests/label-provider-table.stories.ts
@@ -5,7 +5,6 @@ import {
 } from '../../base/tests/label-provider-stories-utils';
 import { labelProviderTableTag } from '..';
 import * as labelTokensNamespace from '../label-tokens';
-import { removePrefixAndCamelCase } from '../../base/tests/label-name-utils';
 
 const metadata = {
     ...labelProviderMetadata,
@@ -18,7 +17,6 @@ export const tableLabelProvider: StoryObj<LabelProviderArgs> = {
     args: {
         labelProviderTag: labelProviderTableTag,
         labelTokens: Object.entries(labelTokensNamespace),
-        prefixSubstring: 'table',
-        removeNamePrefix: removePrefixAndCamelCase
+        prefixSubstring: 'table'
     }
 };